requirements.txt
Chameleon==2.22
Mako==1.0.1
MarkupSafe==0.23
PasteDeploy==1.5.2
Pygments==2.0.2
WebOb==1.4
argparse==1.2.1
py==1.4.26
pyramid==1.5.2
pyramid-chameleon==0.3
pyramid-debugtoolbar==2.3
pyramid-mako==1.0.2
pytest==2.6.4
python-forecastio==1.3.1
repoze.lru==0.6
requests==2.5.1
translationstring==1.3
venusian==1.0
waitress==0.8.9
-e git+git@github.com:jaycee/flightplan-service.git@07649627946df7abdf1a5a49e58dd0e6bcb60ff6#egg=weather_service-master
wsgiref==0.1.2
zope.deprecation==4.1.2
zope.interface==4.1.2
